AudioKit Support and Prepare RP2040

+++ src/AudioCodecs/CodecWAV.h
@@ -46,7 +46,7 @@ class WAVHeader  {
         };
 
         void begin(uint8_t* buffer, size_t len){
-            LOGI("WAVHeader len: %zu", len);
+            LOGI("WAVHeader len: %u",(unsigned) len);
 
             this->buffer = buffer;
             this->len = len;
@@ -212,7 +212,7 @@ class WAVHeader  {
         }
 
         void logInfo(){
-            LOGI("WAVHeader sound_pos: %zu", sound_pos);
+            LOGI("WAVHeader sound_pos: %lu", (unsigned long) sound_pos);
             LOGI("WAVHeader channels: %d ", headerInfo.channels);
             LOGI("WAVHeader bits_per_sample: %d", headerInfo.bits_per_sample);
             LOGI("WAVHeader sample_rate: %d ", headerInfo.sample_rate);
@@ -312,7 +312,7 @@ class WAVDecoder : public AudioDecoder {
                         isValid = header.audioInfo().is_valid;
 
                         LOGI("WAV sample_rate: %d", header.audioInfo().sample_rate);
-                        LOGI("WAV data_length: %zu", header.audioInfo().data_length);
+                        LOGI("WAV data_length: %u", (unsigned) header.audioInfo().data_length);
                         LOGI("WAV is_streamed: %d", header.audioInfo().is_streamed);
                         LOGI("WAV is_valid: %s", header.audioInfo().is_valid ? "true" :  "false");
                         
@@ -433,7 +433,7 @@ class WAVEncoder : public AudioEncoder {
             audioInfo.byte_rate = audioInfo.sample_rate * audioInfo.bits_per_sample * audioInfo.channels;
             audioInfo.block_align =  audioInfo.bits_per_sample / 8 * audioInfo.channels;
             if (audioInfo.is_streamed || audioInfo.data_length==0 || audioInfo.data_length >= 0x7fff0000) {
-                LOGI("is_streamed! because length is %zu", audioInfo.data_length);
+                LOGI("is_streamed! because length is %u",(unsigned) audioInfo.data_length);
                 audioInfo.is_streamed = true;
                 audioInfo.data_length = ~0;
             } else {
